+++ src/WebRoutes.tsx
@@ -56,6 +56,7 @@ import { HighLightNews } from "./pages/NewsPage/HighLightNews";
 import { CreateHighLightNewsPage } from "./pages/NewsPage/CreateHighLightNews";
 import { IndexProductShop } from "./pages/ProductShop/IndexProductShop";
 import { ProductShopDetail } from "./pages/ProductShop/ProductShopDetail";
+import SyncCustomerPage from "./pages/ShopManagementPage/ShopListPage/syncCustomerPage";
 
 interface IRoute {
   path: string;
@@ -287,6 +288,12 @@ export const protectRoutesData: IRoute[] = [
         permission: null,
       },
     ],
+  },
+  {
+    path: "/SyncCustomer",
+    element: <SyncCustomerPage />,
+    permission: null,
+    nestedRoutes: [],
   },
   {
     path: "/UserPage/*",